pipcl.py: various improvements
Use `python-config --ldflags` on unix.

Show wheel contents.

Fix internal names in shared libraries on macos: on MacOS we need to change
mentions of sub-libraries to `@rpath/<leafname>`, otherwise Python will not be
able to import the extension library.

build_extension(): on MacOS, also accept .dylib; assert that only one of .dylib
and .so exist.